Added continuous integration + deployment to GitLab
When a branch is pushed, it builds the weekplanner and tests the unit test project. If it's a master or release branch, it also spits out an .apk file (downloadable through GitHub/pipelines), to be used for testing the app on android or to be put onto the store.

Current issues: unit tests are broken in the master branch, but this issue doesn't reside in gitlab-ci. Why someone ever pushed that to master, I cannot say. Also, while building apk file we turn off AndroidLinkMode. This makes the apk file 60mb instead of 20mb. We turn it off, because the gitlab runners that run the pipeline job don't have enough resources and would otherwise stop responding while building.